refactor: rename compilers/ → renderers/, remove @classmethod, expand exports
- Rename package compilers/ → renderers/ and yaml_compiler.py → yaml_renderer.py
  (package contained only YamlRendererBase, a renderer not a compiler).
- XsdReader: replace @classmethod from_file/from_url with unified __init__
  that accepts file path, URL, or raw XML string.
- Export expand, scan_for_pointers, quick_ref, RenderNode from __init__.py.

608 tests pass, ruff clean, mypy errors reduced by 1.
