+++ src/genro_builders/renderers/__init__.py
@@ -1,6 +1,6 @@
 # Copyright 2025 Softwell S.r.l. - SPDX-License-Identifier: Apache-2.0
 """Built-in renderers for genro-builders."""
 
-from genro_builders.compilers.yaml_compiler import YamlRendererBase
+from genro_builders.renderers.yaml_renderer import YamlRendererBase
 
 __all__ = ["YamlRendererBase"]
